src/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.34.144
botocore==1.34.144
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
coverage==7.6.0
cryptography==42.0.8
daphne==4.1.2
Django==5.0.6
django-after-response==0.2.2
django-allauth==0.63.3
django-autoslug==1.9.9
django-crispy-forms==2.3
django-eventstream==5.1.0
django-grip==3.5.0
django-htmx==1.17.3
django-magiclink==1.3.0
django-markdownify==0.9.5
django-ratelimit==4.1.0
django-storages==1.14.4
django-template-partials==24.2
django-widget-tweaks==1.5.0
elevenlabs==1.3.1
gripcontrol==4.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
jmespath==1.0.1
Markdown==3.6
markdownify==0.12.1
MarkupSafe==2.1.5
msgpack==1.0.8
msgpack-python==0.5.6
ollama==0.2.1
packaging==21.3
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pubcontrol==3.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
tinycss2==1.2.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
xkcdpass==1.19.9
zope.interface==6.4.post2
